prevent possible race conditions while generating invite code
