+++ backend/src/main/java/com/pm4/istp/course/services/CourseInviteCodeHelper.java
@@ -2,31 +2,51 @@ package com.pm4.istp.course.services;
 
 import com.pm4.istp.course.db.entities.Course;
 import com.pm4.istp.course.exceptions.CourseNotFoundException;
+import com.pm4.istp.course.exceptions.InviteCodeGenerationException;
 import com.pm4.istp.course.repositories.CourseRepository;
+import java.security.SecureRandom;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 /**
- * Helper component for invite-code assignment that runs each attempt in its own transaction.
- * Running in {@link Propagation#REQUIRES_NEW} ensures the transaction can be rolled back on a
- * unique-constraint violation without corrupting the caller's session, enabling the caller to retry
- * with a freshly generated code.
+ * Owns all invite-code generation and assignment. Each assignment attempt runs in its own {@link
+ * Propagation#REQUIRES_NEW} transaction so a unique-constraint violation only rolls back that
+ * attempt, leaving the caller's transaction intact for a retry.
  */
 @Component
 @RequiredArgsConstructor
 public class CourseInviteCodeHelper {
 
+  private static final String INVITE_CODE_CHARS = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
+  private static final int INVITE_CODE_LENGTH = 6;
+  private static final SecureRandom SECURE_RANDOM = new SecureRandom();
+
   private final CourseRepository courseRepository;
 
-  /**
-   * Assigns {@code code} to the course identified by {@code courseId} and persists the change
-   * within a new, independent transaction.
-   *
-   * @throws CourseNotFoundException if the course no longer exists
-   */
+  public String generateAndAssign(UUID courseId) {
+    for (int attempt = 0; attempt < 10; attempt++) {
+      String code = generateCode();
+      try {
+        assignInviteCode(courseId, code);
+        return code;
+      } catch (DataIntegrityViolationException ex) {
+        if (!isInviteCodeConstraintViolation(ex)) {
+          throw ex;
+        }
+        if (attempt == 9) {
+          throw new InviteCodeGenerationException(
+              "Could not generate a unique invite code after 10 attempts", ex);
+        }
+      }
+    }
+    throw new InviteCodeGenerationException(
+        "Could not generate a unique invite code after 10 attempts");
+  }
+
   @Transactional(propagation = Propagation.REQUIRES_NEW)
   public void assignInviteCode(UUID courseId, String code) {
     Course course =
@@ -38,5 +58,50 @@ public class CourseInviteCodeHelper {
                         String.format("Course with ID '%s' not found", courseId)));
     course.setInviteCode(code);
     courseRepository.save(course);
+  }
+
+  public Course saveNewCourseWithInviteCode(Course course) {
+    for (int attempt = 0; attempt < 10; attempt++) {
+      String code = generateCode();
+      try {
+        return saveNewCourse(course, code);
+      } catch (DataIntegrityViolationException ex) {
+        if (!isInviteCodeConstraintViolation(ex)) {
+          throw ex;
+        }
+        if (attempt == 9) {
+          throw new InviteCodeGenerationException(
+              "Could not generate a unique invite code after 10 attempts", ex);
+        }
+      }
+    }
+    throw new InviteCodeGenerationException(
+        "Could not generate a unique invite code after 10 attempts");
+  }
+
+  @Transactional(propagation = Propagation.REQUIRES_NEW)
+  public Course saveNewCourse(Course course, String code) {
+    course.setInviteCode(code);
+    return courseRepository.save(course);
+  }
+
+  public String generateCode() {
+    StringBuilder sb = new StringBuilder(INVITE_CODE_LENGTH);
+    for (int i = 0; i < INVITE_CODE_LENGTH; i++) {
+      sb.append(INVITE_CODE_CHARS.charAt(SECURE_RANDOM.nextInt(INVITE_CODE_CHARS.length())));
+    }
+    return sb.toString();
+  }
+
+  private static boolean isInviteCodeConstraintViolation(DataIntegrityViolationException ex) {
+    Throwable current = ex;
+    while (current != null) {
+      String msg = current.getMessage();
+      if (msg != null && msg.contains("uk_courses_invite_code")) {
+        return true;
+      }
+      current = current.getCause();
+    }
+    return false;
   }
 }
