+++ backend/src/main/java/com/pm4/istp/course/services/impl/CourseServiceImpl.java
@@ -19,7 +19,6 @@ import com.pm4.istp.course.exceptions.CourseNotFoundException;
 import com.pm4.istp.course.exceptions.InvalidCourseChallengeException;
 import com.pm4.istp.course.exceptions.InvalidCourseShortDescriptionException;
 import com.pm4.istp.course.exceptions.InvalidInviteCodeException;
-import com.pm4.istp.course.exceptions.InviteCodeGenerationException;
 import com.pm4.istp.course.repositories.ChallengeRepository;
 import com.pm4.istp.course.repositories.CourseEnrollmentRepository;
 import com.pm4.istp.course.repositories.CourseRepository;
@@ -28,7 +27,6 @@ import com.pm4.istp.course.services.CourseService;
 import com.pm4.istp.user.db.entities.User;
 import com.pm4.istp.user.exceptions.UserNotFoundException;
 import com.pm4.istp.user.repositories.UserRepository;
-import java.security.SecureRandom;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Set;
@@ -48,10 +46,6 @@ public class CourseServiceImpl implements CourseService {
 
   private static final String USER_NOT_FOUND_MSG = "User with ID '%s' not found";
   private static final String COURSE_NOT_FOUND_MSG = "Course with ID '%s' not found";
-
-  private static final String INVITE_CODE_CHARS = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
-  private static final int INVITE_CODE_LENGTH = 6;
-  private static final SecureRandom SECURE_RANDOM = new SecureRandom();
 
   private final UserRepository userRepository;
   private final CourseRepository courseRepository;
@@ -77,13 +71,6 @@ public class CourseServiceImpl implements CourseService {
     validateVisibilityState(course.isPublished(), course.isPrivate());
     courseToCreate.setImageUrl(course.getImageUrl());
     courseToCreate.setTopic(course.getTopic());
-    if (course.isPrivate()) {
-      try {
-        courseToCreate.setInviteCode(generateUniqueInviteCode());
-      } catch (IllegalStateException ex) {
-        throw new InviteCodeGenerationException("Unable to generate a unique invite code", ex);
-      }
-    }
 
     // Owner = the user making the request
     CourseInstructor owner = new CourseInstructor();
@@ -112,6 +99,9 @@ public class CourseServiceImpl implements CourseService {
       }
     }
 
+    if (course.isPrivate()) {
+      return courseInviteCodeHelper.saveNewCourseWithInviteCode(courseToCreate);
+    }
     return courseRepository.save(courseToCreate);
   }
 
@@ -208,13 +198,8 @@ public class CourseServiceImpl implements CourseService {
     boolean willBePublished = request.isPublished();
     boolean willBePrivate = request.isPrivate();
     validateVisibilityState(willBePublished, willBePrivate);
-    if (willBePrivate && (!wasPrivate || course.getInviteCode() == null)) {
-      try {
-        course.setInviteCode(generateUniqueInviteCode());
-      } catch (IllegalStateException ex) {
-        throw new InviteCodeGenerationException("Failed to generate invite code", ex);
-      }
-    } else if (!willBePrivate) {
+    boolean needsNewCode = willBePrivate && (!wasPrivate || course.getInviteCode() == null);
+    if (!willBePrivate) {
       course.setInviteCode(null);
     }
     course.setPublished(willBePublished);
@@ -260,7 +245,11 @@ public class CourseServiceImpl implements CourseService {
       }
     }
 
-    return courseRepository.save(course);
+    Course saved = courseRepository.save(course);
+    if (needsNewCode) {
+      saved.setInviteCode(courseInviteCodeHelper.generateAndAssign(saved.getId()));
+    }
+    return saved;
   }
 
   @Override
@@ -399,38 +388,8 @@ public class CourseServiceImpl implements CourseService {
               "Course '%s' is not private; invite code regeneration is disabled", courseId));
     }
 
-    for (int attempt = 0; attempt < 10; attempt++) {
-      String generatedCode = generateInviteCode();
-      try {
-        courseInviteCodeHelper.assignInviteCode(courseId, generatedCode);
-        // Keep response mapping on the already-loaded aggregate to avoid detached lazy
-        // collections.
-        course.setInviteCode(generatedCode);
-        return course;
-      } catch (DataIntegrityViolationException ex) {
-        if (!isInviteCodeConstraintViolation(ex)) {
-          throw ex;
-        }
-        if (attempt == 9) {
-          throw new InviteCodeGenerationException(
-              "Could not generate a unique invite code after 10 attempts", ex);
-        }
-      }
-    }
-    throw new InviteCodeGenerationException(
-        "Could not generate a unique invite code after 10 attempts");
-  }
-
-  private static boolean isInviteCodeConstraintViolation(DataIntegrityViolationException ex) {
-    Throwable current = ex;
-    while (current != null) {
-      String msg = current.getMessage();
-      if (msg != null && msg.contains("uk_courses_invite_code")) {
-        return true;
-      }
-      current = current.getCause();
-    }
-    return false;
+    course.setInviteCode(courseInviteCodeHelper.generateAndAssign(courseId));
+    return course;
   }
 
   private void verifyOwner(Course course, UUID userId) {
@@ -458,24 +417,6 @@ public class CourseServiceImpl implements CourseService {
           String.format(
               "User with ID '%s' is not an instructor of course '%s'", userId, course.getId()));
     }
-  }
-
-  private String generateInviteCode() {
-    StringBuilder sb = new StringBuilder(INVITE_CODE_LENGTH);
-    for (int i = 0; i < INVITE_CODE_LENGTH; i++) {
-      sb.append(INVITE_CODE_CHARS.charAt(SECURE_RANDOM.nextInt(INVITE_CODE_CHARS.length())));
-    }
-    return sb.toString();
-  }
-
-  private String generateUniqueInviteCode() {
-    for (int attempt = 0; attempt < 10; attempt++) {
-      String code = generateInviteCode();
-      if (!courseRepository.existsByInviteCode(code)) {
-        return code;
-      }
-    }
-    throw new IllegalStateException("Could not generate a unique invite code after 10 attempts");
   }
 
   private String normalizeShortDescription(String shortDescription) {
